Addressing last round of review comments
